website/requirements.txt
Django==1.4.1
South==0.7.6
argparse==1.2.1
django-crispy-forms==1.1.4
django-registration==0.8
django-social-auth==0.7.4
django-tastypie==0.9.11
gondor==1.1.3
gunicorn==0.14.6
httplib2==0.7.4
mimeparse==0.1.3
oauth2==1.5.211
psycopg2==2.4.5
python-dateutil==1.5
python-openid==2.2.5
simplejson==2.6.1
wsgiref==0.1.2
